[Tests] Split runners from impl for TD integrations tests
