+++ integration-test/trace-debugger/src/main/org/jetbrains/trace/debugger/test/impl/TraceDebuggerExamplesTraceDebuggerIntegrationTest.kt
@@ -8,8 +8,9 @@
  * with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package org.jetbrains.trace.debugger.test
+package org.jetbrains.trace.debugger.test.impl
 
+import org.jetbrains.trace.debugger.test.runner.AbstractTraceDebuggerIntegrationTest
 import org.junit.Ignore
 import org.junit.Test
 import java.nio.file.Paths
